[ADD]ExperimentService中增加了updateCommitNumber方法
[update]修改了CourseController的list方法中关于Byte的相等判断
[update]增加了RecordController中保存方法的更新提交人数操作

+++ src/main/java/com/gdou/teaching/controller/student/StudentCourseController.java
@@ -123,14 +123,14 @@ public class StudentCourseController {
             List<CourseDTO> list = courseService.list(user.getUserId());
             //拿到了数据，进行分割，将数据分为"未结束"和"已结束"两部分
             //切割正常课程。
-            List<CourseVO> normal = list.stream().filter(c->c.getCourseStatus().equals(CourseStatusEnum.NORMAL.getCode()))
+            List<CourseVO> normal = list.stream().filter(c->c.getCourseStatus().equals(CourseStatusEnum.NORMAL.getCode().byteValue()))
                     .map(courseDTO -> {
                 CourseVO courseVO = new CourseVO();
                 BeanUtils.copyProperties(courseDTO, courseVO);
                 return courseVO;
             }).collect(Collectors.toList());
             //切割出过期课程
-            List<CourseVO> end = list.stream().filter(c->c.getCourseStatus().equals(CourseStatusEnum.END.getCode())).map(courseDTO -> {
+            List<CourseVO> end = list.stream().filter(c->c.getCourseStatus().equals(CourseStatusEnum.END.getCode().byteValue())).map(courseDTO -> {
                 CourseVO courseVO = new CourseVO();
                 BeanUtils.copyProperties(courseDTO, courseVO);
                 return courseVO;
